fix postgres

+++ pages/api/todos/[id].ts
@@ -5,10 +5,13 @@ export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
-  const { id } = req.query;
+  if (!req.query.id) {
+    return res.status(400).json({ error: 'ID is required' });
+  }
 
   try {
     if (req.method === 'DELETE') {
+      const id = Array.isArray(req.query.id) ? req.query.id[0] : req.query.id;
       await sql`DELETE FROM todos WHERE id = ${Number(id)};`;
       return res.status(200).json({ message: 'Todo deleted successfully' });
     }
